Refactor the integration test files
