+++ src/tests/api/v1/utils.ts
@@ -1,0 +1,9 @@
+export { ADMIN_SECRET } from '../../../lib/config';
+
+export const BASE_URL = '/api/v1';
+
+export const USERS_URL = `${BASE_URL}/users`;
+
+export const POSTS_URL = `${BASE_URL}/posts`;
+
+export const SIGNIN_URL = `${BASE_URL}/auth/signin`;
